DE434636: Store lastUuid for each folder in AsmReader
In case of restart of the threads, keep the previous value of uuid in order to continue from that point so that we can always have the most recent data. This might produce some small gaps after each restart in case there is no new data after last query time.